codex(refactor): slim compose orchestrator
